cloud_functions/start_gmail_watcher/requirements.txt
blinker==1.7.0
cachetools==5.3.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
cloudevents==1.10.1
colorama==0.4.6
deprecation==2.1.0
exceptiongroup==1.2.0
Flask==3.0.2
functions-framework==3.5.0
google-api-core==2.17.1
google-api-python-client==2.118.0
google-auth==2.28.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
google-cloud-secret-manager==2.18.2
googleapis-common-protos==1.62.0
greenlet==3.0.3
grpc-google-iam-v1==0.13.0
grpcio==1.62.0
grpcio-status==1.62.0
httplib2==0.22.0
idna==3.6
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
oauthlib==3.2.2
packaging==23.2
pluggy==1.4.0
proto-plus==1.23.0
protobuf==4.25.3
psycopg2-binary==2.9.9
pyasn1==0.5.1
pyasn1-modules==0.3.0
pyparsing==3.1.1
pytest==8.0.2
python-dotenv==1.0.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
SQLAlchemy==1.4.46
tomli==2.0.1
typing_extensions==4.10.0
uritemplate==4.1.1
urllib3==2.2.1
watchdog==4.0.0
Werkzeug==3.0.1
